+++ CHANGES.rst
@@ -5,6 +5,7 @@ LATEST
 ------
 - Converted widgets to a sub-package.
 - Fixed issue with labels in a layout column preventing buttons from being pressed.
+- Fixed issue with visual overrun on Listboxes when there is a label offset.
 
 1.12.0
 ------