+++ CHANGES.rst
@@ -4,7 +4,7 @@ CHANGE HISTORY
 LATEST
 ------
 - Added ability to change a `Button`'s text through a `.text` attribute.
-- Added ability to accept a name attribute in the `Button` constructor.
+- Added ability to accept a name attribute in the `Button` and `Label` constructors.
 - Added ability to detect job pause/resume and force full screen refresh.
 - Added ability to request terminal default colours using `Screen.COLOUR_DEFAULT`.
 - Converted widgets to a sub-package.